--- ext/simplexml/sxe.h
@@ -1,27 +1,0 @@
-/*
-   +----------------------------------------------------------------------+
-   | Copyright (c) The PHP Group                                          |
-   +----------------------------------------------------------------------+
-   | This source file is subject to version 3.01 of the PHP license,      |
-   | that is bundled with this package in the file LICENSE, and is        |
-   | available through the world-wide-web at the following url:           |
-   | http://www.php.net/license/3_01.txt                                  |
-   | If you did not receive a copy of the PHP license and are unable to   |
-   | obtain it through the world-wide-web, please send a note to          |
-   | license@php.net so we can mail you a copy immediately.               |
-   +----------------------------------------------------------------------+
-   | Authors: Marcus Boerger <helly@php.net>                              |
-   +----------------------------------------------------------------------+
- */
-
-#ifndef SXE_H
-#define SXE_H
-
-#include "php.h"
-
-extern PHP_SXE_API zend_class_entry *ce_SimpleXMLIterator;
-extern PHP_SXE_API zend_class_entry *ce_SimpleXMLElement;
-
-PHP_MINIT_FUNCTION(sxe);
-
-#endif /* SXE_H */
