--- ext/simplexml/sxe.stub.php
@@ -1,27 +1,0 @@
-<?php
-
-/** @generate-function-entries */
-
-class SimpleXMLIterator
-{
-    /** @return void */
-    public function rewind() {}
-
-    /** @return bool */
-    public function valid() {}
-
-    /** @return SimpleXMLElement|null */
-    public function current() {}
-
-    /** @return string|false */
-    public function key() {}
-
-    /** @return void */
-    public function next() {}
-
-    /** @return bool */
-    public function hasChildren() {}
-
-    /** @return SimpleXMLIterator|null */
-    public function getChildren() {}
-}
