--- ext/simplexml/sxe_arginfo.h
@@ -1,37 +1,0 @@
-/* This is a generated file, edit the .stub.php file instead. */
-
-ZEND_BEGIN_ARG_INFO_EX(arginfo_class_SimpleXMLIterator_rewind, 0, 0, 0)
-ZEND_END_ARG_INFO()
-
-#define arginfo_class_SimpleXMLIterator_valid arginfo_class_SimpleXMLIterator_rewind
-
-#define arginfo_class_SimpleXMLIterator_current arginfo_class_SimpleXMLIterator_rewind
-
-#define arginfo_class_SimpleXMLIterator_key arginfo_class_SimpleXMLIterator_rewind
-
-#define arginfo_class_SimpleXMLIterator_next arginfo_class_SimpleXMLIterator_rewind
-
-#define arginfo_class_SimpleXMLIterator_hasChildren arginfo_class_SimpleXMLIterator_rewind
-
-#define arginfo_class_SimpleXMLIterator_getChildren arginfo_class_SimpleXMLIterator_rewind
-
-
-ZEND_METHOD(SimpleXMLIterator, rewind);
-ZEND_METHOD(SimpleXMLIterator, valid);
-ZEND_METHOD(SimpleXMLIterator, current);
-ZEND_METHOD(SimpleXMLIterator, key);
-ZEND_METHOD(SimpleXMLIterator, next);
-ZEND_METHOD(SimpleXMLIterator, hasChildren);
-ZEND_METHOD(SimpleXMLIterator, getChildren);
-
-
-static const zend_function_entry class_SimpleXMLIterator_methods[] = {
-	ZEND_ME(SimpleXMLIterator, rewind, arginfo_class_SimpleXMLIterator_rewind, ZEND_ACC_PUBLIC)
-	ZEND_ME(SimpleXMLIterator, valid, arginfo_class_SimpleXMLIterator_valid, ZEND_ACC_PUBLIC)
-	ZEND_ME(SimpleXMLIterator, current, arginfo_class_SimpleXMLIterator_current, ZEND_ACC_PUBLIC)
-	ZEND_ME(SimpleXMLIterator, key, arginfo_class_SimpleXMLIterator_key, ZEND_ACC_PUBLIC)
-	ZEND_ME(SimpleXMLIterator, next, arginfo_class_SimpleXMLIterator_next, ZEND_ACC_PUBLIC)
-	ZEND_ME(SimpleXMLIterator, hasChildren, arginfo_class_SimpleXMLIterator_hasChildren, ZEND_ACC_PUBLIC)
-	ZEND_ME(SimpleXMLIterator, getChildren, arginfo_class_SimpleXMLIterator_getChildren, ZEND_ACC_PUBLIC)
-	ZEND_FE_END
-};
